pyproject.toml
[project]
name = "clockchain-prediction-market"
version = "0.1.0"
description = "Decentralized prediction market on BASE blockchain"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.3",
    "cryptography>=45.0.5",
    "email-validator>=2.2.0",
    "flask-limiter>=3.12",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-levenshtein>=0.27.1",
    "pytz>=2025.2",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "web3>=7.12.1",
    "websockets>=15.0.1",
    "werkzeug>=3.1.3",
    "playwright>=1.54.0",
    "pillow>=11.3.0",
    "tweepy>=4.16.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["services", "routes"]
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
